refactor: move shared fn to a separate file
